// frontend/lib/generated/fastapi/schemas/homepageFeedItem.ts
/**
 * Generated by orval v7.20.0 🍺
 * Do not edit manually.
 * ShipyardHQ API
 * OpenAPI spec version: 0.1.0
 */
import type { HomepageFeedItemCategory } from './homepageFeedItemCategory';
import type { HomepageFeedItemCategorySlug } from './homepageFeedItemCategorySlug';
import type { HomepageFeedItemScoreCount } from './homepageFeedItemScoreCount';
import type { HomepageFeedItemUpdatesCount } from './homepageFeedItemUpdatesCount';
import type { HomepageFeedItemVariant } from './homepageFeedItemVariant';
import type { HomepageFeedItemLatestRevenueCents } from './homepageFeedItemLatestRevenueCents';
import type { HomepageFeedItemRevenueCurrencyCode } from './homepageFeedItemRevenueCurrencyCode';
import type { HomepageFeedItemInterest } from './homepageFeedItemInterest';

export interface HomepageFeedItem {
  id: string;
  slug: string;
  name: string;
  logo: string;
  tagline: string;
  createdAt: string;
  updatedAt: string;
  badges: string[];
  category?: HomepageFeedItemCategory;
  categorySlug?: HomepageFeedItemCategorySlug;
  scoreCount?: HomepageFeedItemScoreCount;
  updatesCount?: HomepageFeedItemUpdatesCount;
  isSponsored: boolean;
  isVoted: boolean;
  isVerified: boolean;
  variant?: HomepageFeedItemVariant;
  latestRevenueCents?: HomepageFeedItemLatestRevenueCents;
  revenueCurrencyCode?: HomepageFeedItemRevenueCurrencyCode;
  interest?: HomepageFeedItemInterest;
  shuffleRank: number;
}